:bug: Separate control from GCC & e-except
GCC needs more memory because it needs to allocate the exception
header for every exception.
